+++ packages/chakra-ui-core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 0.6.5
+
+### Patch Changes
+
+- fix style props for @chakra-ui/vue
+
+## 0.6.4
+
+### Patch Changes
+
+- improvements to style props
+
+## 0.6.3
+
+### Patch Changes
+
+- fix: includes missing style props for border-radius and flex-grow style declarations"
+
+## 0.6.2
+
+### Patch Changes
+
+- 3533e00: Fix: CSwitch component v-model + @change event handler.
+
+## 0.6.2-next.0
+
+### Patch Changes
+
+- Fix: CSwitch component v-model + @change event handler.
+
 ## 0.6.1
 
 ### Patch Changes