+++ packages/chakra-ui-nuxt/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 0.0.13
+
+### Patch Changes
+
+- fix style props for @chakra-ui/vue
+- Updated dependencies [undefined]
+  - @chakra-ui/vue@0.6.5
+
+## 0.0.12
+
+### Patch Changes
+
+- improve component style props
+
+## 0.0.10
+
+### Patch Changes
+
+- 3533e00: Fix: CSwitch component v-model + @change event handler.
+- Updated dependencies [3533e00]
+  - @chakra-ui/vue@0.6.2
+
+## 0.0.10-next.0
+
+### Patch Changes
+
+- Fix: CSwitch component v-model + @change event handler.
+- Updated dependencies [undefined]
+  - @chakra-ui/vue@0.6.2-next.0
+
 ## 0.0.7
 
 ### Patch Changes